+++ src/js/index.js
@@ -16,3 +16,17 @@ async function allTest() {
 
 
 allTest();
+
+
+
+console.log(window.location.pathname)
+//have to put in extra here for github pages
+if (window.location.pathname == "/index.html" || window.location.pathname == "/" || window.location.pathname == "Auction-House/index.html" || window.location.pathname == "Auction-House/") {
+    const json = await calls.allListings()
+    console.log(json)
+
+let container = document.querySelector("#container")
+
+utils.postMaker(json)
+    
+}
